+++ frontend/demo/component/login/login-overlay-mockup.ts
@@ -3,7 +3,7 @@ import { css, html, LitElement, unsafeCSS } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import type { LoginI18n } from '@vaadin/login';
 import { applyTheme } from 'Frontend/generated/theme';
-import img from '../../../../src/main/resources/images/starry-sky.png';
+import img from '../../../../src/main/resources/images/starry-sky.png?url';
 
 @customElement('login-overlay-mockup')
 export class LoginOverlayMockupElement extends LitElement {
