+++ frontend/demo/component/scroller/scroller-both.ts
@@ -3,7 +3,7 @@ import '@vaadin/scroller';
 import { html, LitElement } from 'lit';
 import { customElement } from 'lit/decorators.js';
 import { applyTheme } from 'Frontend/generated/theme';
-import img from '../../../../src/main/resources/images/reindeer.jpg';
+import img from '../../../../src/main/resources/images/reindeer.jpg?url';
 
 @customElement('scroller-both')
 export class Example extends LitElement {
